fix pir ut test_adamw_op_(xpu)
