--- test/deprecated/legacy_test/test_adamw_op_deprecated.py
@@ -1,61 +1,0 @@
-# Copyright (c) 2020 PaddlePaddle Authors. All Rights Reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-import unittest
-
-import numpy as np
-
-import paddle
-from paddle import base
-
-
-class TestAdamWOp(unittest.TestCase):
-    def test_adamw_op(self):
-        paddle.enable_static()
-        place = base.CPUPlace()
-        shape = [2, 3, 8, 8]
-        exe = base.Executor(place)
-        train_prog = base.Program()
-        startup = base.Program()
-        with base.program_guard(train_prog, startup):
-            with base.unique_name.guard():
-                data = paddle.static.data(name="data", shape=shape)
-                conv = paddle.static.nn.conv2d(data, 8, 3)
-                loss = paddle.mean(conv)
-
-                beta1 = paddle.static.create_global_var(
-                    shape=[1], value=0.85, dtype='float32', persistable=True
-                )
-                beta2 = paddle.static.create_global_var(
-                    shape=[1], value=0.95, dtype='float32', persistable=True
-                )
-                betas = [beta1, beta2]
-                opt = paddle.optimizer.AdamW(
-                    learning_rate=1e-5,
-                    beta1=beta1,
-                    beta2=beta2,
-                    weight_decay=0.01,
-                    epsilon=1e-8,
-                )
-                opt.minimize(loss)
-
-        exe.run(startup)
-        data_np = np.random.random(shape).astype('float32')
-        rets = exe.run(train_prog, feed={"data": data_np}, fetch_list=[loss])
-        assert rets[0] is not None
-        paddle.disable_static()
-
-
-if __name__ == "__main__":
-    unittest.main()
